Jai API

Using Jai API, Julia user can easily run CPU or GPU codes written in Fortran/C/C++.

list of Jai API

Jai API is consist of serveral Julia macros.

AccelInterfaces.@jaccel

AccelInterfaces.@jdecel

AccelInterfaces.@jenterdata

AccelInterfaces.@jexitdata

AccelInterfaces.@jkernel

AccelInterfaces.@jlaunch

AccelInterfaces.@jwait

Detailed API

AccelInterfaces.@jaccel — Macro

@jaccel [name][, clauses...]

Create accelerator context.

If name is not specified, this context can be accessed only as the currently active context.

Arguments

name: a unique name for this accelerator context

framework: a list the framework names and compiler-command for the frameworks

device: device number

constant: list of constant variables to be available in all kernels

set: set several options for the accelerator context including "workdir" for working directory, "debug" for enabling debugging feature.

See also @jdecel, @jkernel

Examples

julia> @jaccel myacc framework(fortran="gfortran -fPIC -shared -g")
AccelInfo

AccelInterfaces.@jdecel — Macro

@jdecel [name]

Delete an accelerator context.

If name is not specified, this context can be accessed only as the currently active context.

Arguments

name: a unique name for this accelerator context

See also @jaccel, @jkernel

Examples

julia> @jdecel myacc

AccelInterfaces.@jenterdata — Macro

@jenterdata [name][, clauses...]

Allocate device memory or copy data to device memory.

If name is not specified, the currently active accel context will be used.

Arguments

name: a unique name for the accelerator context

alloc: a list of variable names to be allocated in an accelerator memory

updateto: a list of variable names whose content will be copied to an accelerator

See also @jaccel, @jkernel

Examples

julia> @jenterdata myacc alloc(X, Y, Z) updateto(X, Y, Z)

AccelInterfaces.@jexitdata — Macro

@jexitdata [name][, clauses...]

Dealloc device memory or copy data from device memory.

If name is not specified, the currently active accel context will be used.

Arguments

name: a unique name for the accelerator context

delete: a list of variable names to be deallocated in an accelerator memory

updatefrom: a list of variable names whose GPU content will be copied to host

See also @jaccel, @jkernel

Examples

julia> @jexitdata myacc delete(X, Y, Z), updatefrom(X, Y, Z)

AccelInterfaces.@jkernel — Macro

@jkernel kerneldef, [kernelname, [accelname, ]][clauses...]

Create kernel context.

If kernelname or accelname is not specified, the currently active accel or kernel context will be used.

Arguments

kerneldef: Jai kernel definition

kernelname: Kernel context name

accelname: Accelerator context name

framework: a list the framework names and compiler-command for the frameworks

See also @jaccel, @jenterdata

Examples

julia> @jkernel filepath mykernel myaccel framework(fortran="gfortran -fPIC -shared -g")

AccelInterfaces.@jlaunch — Macro

@jlaunch [kernelname, [accelname, ]][clauses...]

Launch a kernel on an accelerator.

If kernelname or accelname is not specified, the currently active accel or kernel context will be used.

Arguments

kernelname: Kernel context name

accelname: Accel context name

input: a list of input variables to the kernel

output: a list of output variables to the kernel

hip: a list of hip launch configurations

cuda: a list of cuda launch configurations

See also @jaccel, @jkernel

Examples

julia> @jlaunch mykernel myaccel input(X, Y) output(Z) hip(threads=((1,1,1), (256,1,1)),enable_if=true)

AccelInterfaces.@jwait — Macro

@jwait [accelname]

Wait to finish device operation

If accelname is not specified, the currently active accel context will be used.

Arguments

accelname: Accel context name

See also @jaccel, @jkernel

Examples

julia> @jwait myaccel